extract typechecker code from ExamplesTest
